linux failure for custom script?

+++ src/server/scripts/Custom/System_OnKill.cpp
@@ -16,7 +16,7 @@ class System_OnKill : public PlayerScript
 			uint64 lastkill;
 		};
 
-		static std::map<uint64, KillStreak_Info> KillStreakData;
+		static std::map<uint32, KillStreak_Info> KillStreakData;
 
 		if( KillerGUID == VictimGUID || KillStreakData[KillerGUID].lastkill == VictimGUID )
 			return;
